backend small refactoring https://github.com/TcoiOleg/DisciplineApp/issues/29
