+++ src/main/backend/java/com/olts/discipline/rest/api/UserRestController.java
@@ -4,8 +4,8 @@ import com.olts.discipline.api.service.HabitService;
 import com.olts.discipline.api.service.UserService;
 import com.olts.discipline.entity.Habit;
 import com.olts.discipline.entity.User;
-import com.olts.discipline.rest.assembler.PageableResourceAssembler;
-import com.olts.discipline.rest.dto.PageableResource;
+import com.olts.discipline.rest.hateoas.assembler.PageableResourceAssembler;
+import com.olts.discipline.rest.hateoas.PageableResource;
 import com.olts.discipline.rest.dto.UserGETDto;
 import com.olts.discipline.rest.dto.UserPutDto;
 import com.olts.discipline.rest.mapper.HabitMapper;
