+++ src/main/backend/java/com/olts/discipline/rest/hateoas/PageDto.java
@@ -1,4 +1,4 @@
-package com.olts.discipline.rest.dto;
+package com.olts.discipline.rest.hateoas;
 
 import lombok.Data;
